requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
python-dotenv==1.2.1
requests==2.32.5
pydantic==2.12.5
groq==0.36.0